Sun radiation can be computed using orbit latitude
The function to compute the sun radiation can propagate the orbit using the
orbit latitude or the mean anomaly. Notice that if the latter is used, then
the average value of the output will be the average sun radiation that the
surface will receive, because the angular steps have the same time
interval.