:sparkles: Add function to compute the Equation of Time
It is now possible to compute the equation of time, which is the
difference between the Solar Apparent Local Time and the Solar Mean
Local Time. This was necessary to correct the errors that were seen
when computing the RAAN.

In this commit, the function to compute RAAN was also updated. This
was a necessary step for FOrPlan/FOrPlan#29.