:construction: :wrench: Change SGP4 update function interface
The function `sgp4` was renamed to `sgp4!` to keep naming
convention of Julia, since it changes values of its arguments.
Moreover, it now returns only the position and velocity vector at
the time requested. The mean orbital elements are updated inside
the structure `sgp4d`.

Notice that this is still subject to change.