:warning: :wrench: Improve computation of legendre functions
It is now possible to select different normalization (fully
normalization or Schmidt quasi-normalization) and it is also
possible to add the Condon-Shortley phase term to the computations.

This is a **BREAKING COMMIT**. Because the functions `legendre` and
`legendre!` now computes `P_n,m[cos(ϕ)]` instead of
`P_n,m[sin(ϕ)]`.