:construction: IGRF v12 Model
The initial version of IGRF v12 model written in julia language is
available in this commit. Two new functions are added:

    - `igrf12`
    - `igrf12syn`

The first is an independent implementation of the IGRF with a more
readable code. The second is the original FORTRAN code converted to
julia, with the same inputs and outputs.

The `igrf12` currently supports only the geocentric reference
frame. Hence, this is still a work in progress.