:rotating_light: Fix incompatibilities with AppVeyor / Windows
`git clone` was failing on Windows because the directory called
`aux`. This is a reserved word in windows and cannot be used.
Hence, `./src/aux` was changed to `./src/misc`.

For the same reason, the file `transformations/aux.jl` was renamed
to `transformations/misc.jl`.

This commit also added the AppVeyor badge to `README.md`.