:bug: Workaround to avoid problem at singularity
It is better to use the `cos` function to compute the cosine of the
latitude in the gravity model to avoid problems at singularity. In
this case, if the latitude is 90 deg, then `cos` will return a very
small number but not 0. Hence, the second coordinate of the
acceleration represented in the local reference frame will not be
`NaN`.