:sparkles: Add Jacchia-Roberts atmosphere model
The Jacchia-Roberts 1971 atmosphere model was added to the toolbox.
This is an analytic version of Jacchia 1971.

In preliminary tests, this model can compute the atmospheric
density 10x faster than NRLMSISE-00 in most of cases. However, the
accuracy is expected to be lower.

The implementation of this model was requested in issue #15.