:wrench: Add generic parse to ICGEM files
The ICGEM files can now be parsed using the function `parse_icgem`,
which returns an instance of the structure `ICGEM`. This
information can be used to compute the gravity by creating the
coefficients using the function `create_gravity_model_coefs`.

With this commit, the old function `parse_gfc(filename)` is
deprecated in favor of
`create_gravity_model_coefs(parse_icgem(filename))`.

Closes #17