:wrench: Simplify the API of the frame transformations
It turns out that the model used to compute the coordinate frame
rotations can be inferred only by looking at the selected frames.
This is true because the IAU2000 theory will be implemented based
on the CIO approach, which has different intermediate frames.

All the old style calls, which has the model information, is marked
as deprecated now.

Closes #18