:bug: :warning: Fix EOP data download
For some reason, the HTTP request to download the EOP data started
to being forbidden. Hence, it was decided to use the package
`RemoteFiles` to download the files and store on the file system
before they are parsed.

This was the easiest / quickest fix possible to the aforementioned
problem. In fact, the real fix to this kind of problems is
described in issue #2.

Notice that this can be breaking because now the file is downloaded
and stored in the file system instead of being read from the
memory. However, it is highly unlikely that this commit will break
existing code.