:warning: Remove deprecations related to frame rotations
All the deprecations warnings related to frame rotations, which
were introduced in v0.5, were removed. This was necessary because
it can interfere with the new IAU-2006 theory added. For example,
calling:

    rECEFtoECI(FK5(), GCRF(), CIRS(), eop)

with `eop` obtained from IAU2000A, would compute the rotation from
GCRF to CIRS using the IAU-2006, even though the user explicit
asked for the FK5 theory. To avoid such confusion, all the
functions call that have the model as a parameter will now throw an
error.