:bug: :warning: Fix bug in J2 orbit propagator
When using the API to initialize the J2 orbit propagator, the
parameter `n_0` is the mean motion, i.e. the mean angular velocity
between two consecutive passages to the perigee. The algorithm
obtained from Vallado's book, on the other hand, seems to consider
that `n_0` is the rate of change of the mean anomaly. Hence, we
must subtract the time-derivative of the perigee in order to make
the algorithms compatible.

Some variables were also cached into the J2 structure. Hence, the
algorithm performance increase by 25% during orbit propagation.