:warning: :construction: Improve propagators API - J2
J2 initialization function `j2_init` now receives the 6 classical
orbital elements as the input. Those are considered to be the
inital mean elements.

The API `init_orbit_propagator` related to J2 was also changed to
receive the 6 classical orbital elements.

In case a TLE is used to initialize the J2 propagator, the
semi-major axis is recovered using the same algorithm in SGP4.