:warning: Simplify dlegendre API
The API function in which only `P` is passed was removed from all
functions related to `dlegrendre`. If the user has the matrix `P`,
then it must allocate the matrix `dP` and call `dlegendre`.
However, the function `dlegrendre` that allocates everything now
returns both the matrix `dP` and the matrix `P`. Thus, this is a
breaking change.