:sparkles: :warning: Select maximum order in grav. models
In all functions related to the gravity models, it is now possible
to select the maximum degree and order. However, there was a
breaking change in the API. Selecting `n_max` 0 will lead to the
same result as `n_max` 1. This was selected to maintain consistency
with the order `m_max` that can be 0. If the user desired to
compute using all the coefficientes, then `n_max` and `m_max` must
be set to -1, which is the default.