:wrench: Improve function to compute eclipse time
The function that computes the satellite eclipse time was renamed
from `satellite_eclipse_time` to `eclipse_time_summary` (the old
one is now marked as deprecated).

The function was drastically improved by adding a border finding
algorithm to precisely compute the time in which the zones start
(sunlight, penumbra, and umbra). Hence, it was possible to increase
the default time step and also increase the accuracy. This new
version is almost 3x faster than the old one.