:warning: Use separated type for epochs in propagators
Usually, the epoch in the propagators is the Julian Day in which the
initial mean elements were defined. A Julian Day should be represented
using a 64-bit floats. However, the user could select 32-bit floats for
internal variables to speed up computations in the propagator. Hence, we
need to split the type at every propagator between the type for the
epoch and the type for the internal variables.