Clear confusion about geomagnetic field vector
The input variables to these functions are date (i.e. year), radius (i.e. Z), latitude, and longitude (i.e. X and Y). There's no mention that the components of the returned geomagnetic field vector are X, Y, Z. While this might be an obvious convention, I had no idea and assumed that the order of the components would be the same as the order of the input variables to these functions (i.e. Z, X, Y). I tried to "fix" that with that tiny notation in the parenthesis, but maybe you have a better idea.

As a side note, it might be good to add functions for the magnetic components listed for example [here](https://www.geomag.nrcan.gc.ca/mag_fld/comp-en.php).